programs/requirements.txt
asttokens==2.4.0
blas==1.0
bottleneck==1.3.5
brotli==1.0.9
brotli-bin==1.0.9
brotlipy==0.7.0
bzip2==1.0.8
ca-certificates==2023.08.22
caffeine==0.5
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
colorama==0.4.6
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
executing==2.0.0
ffmpeg==4.2.2
filelock==3.9.0
fonttools==4.25.0
freetype==2.12.1
gettext==0.21.0
giflib==5.2.1
gmp==6.2.1
gmpy2==2.1.2
gnutls==3.6.15
icecream==2.1.3
icu==73.1
idna==3.4
importlib_resources==5.2.0
jinja2==3.1.2
joblib==1.2.0
jpeg==9e
kiwisolver==1.4.4
lame==3.100
lcms2==2.12
lerc==3.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcxx==14.0.6
libdeflate==1.17
libffi==3.4.4
libgfortran==5.0.0
libgfortran5==11.3.0
libiconv==1.16
libidn2==2.3.4
libopenblas==0.3.21
libopus==1.3
libpng==1.6.39
libtasn1==4.19.0
libtiff==4.5.1
libunistring==0.9.10
libvpx==1.10.0
libwebp==1.3.2
libwebp-base==1.3.2
libxml2==2.10.4
llvm-openmp==14.0.6
lz4-c==1.9.4
markupsafe==2.1.1
matplotlib==3.7.2
matplotlib-base==3.7.2
mpc==1.1.0
mpfr==4.0.2
mpmath==1.3.0
munkres==1.1.4
ncurses==6.4
nettle==3.7.3
networkx==3.1
numexpr==2.8.4
numpy==1.25.2
numpy-base==1.25.2
openh264==1.8.0
openssl==3.0.11
packaging==23.1
pandas==2.0.3
pillow==9.4.0
pip==23.2.1
pycparser==2.21
pygments==2.16.1
pyopenssl==23.2.0
pyparsing==3.0.9
pysocks==1.7.1
python==3.9.18
python-dateutil==2.8.2
python-tzdata==2023.3
pytorch==2.0.1
pytz==2023.3.post1
readline==8.2
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
setuptools==68.0.0
six==1.16.0
sqlite==3.41.2
sympy==1.11.1
threadpoolctl==2.2.0
tk==8.6.12
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023c
urllib3==1.26.16
wheel==0.38.4
x264==1!152.20180806
xz==5.4.2
zipp==3.11.0
zlib==1.2.13
zstd==1.5.5